Исправление замечаний. Фетч
