+++ js/main.js
@@ -2,4 +2,5 @@ import './popup.js';
 import './form.js';
 import './api.js';
 import './map.js';
+import './filter.js';
